Move controller test to proper package
