+++ android-app/app/src/test/java/tallerii/stories/controller/LoginControllerTest.java
@@ -1,4 +1,4 @@
-package tallerii.stories;
+package tallerii.stories.controller;
 
 import android.widget.Toast;
 
@@ -15,6 +15,7 @@ import java.time.Instant;
 
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
+import tallerii.stories.LoginActivity;
 import tallerii.stories.controller.LoginController;
 import tallerii.stories.network.ConstantsApplicationApiRest;
 import tallerii.stories.network.apimodels.EntityMetadata;
